% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{create_client}
\alias{create_client}
\title{Creates a client record including basic demographics}
\usage{
create_client()
}
\value{
a single row tibble containing client demographics
}
\description{
Creates a client record including basic demographics
}
\examples{
purrr::rerun(20, create_client()) \%>\%
do.call(rbind.data.frame, .)
}
