% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{create_contact}
\alias{create_contact}
\title{Creates a contact episode}
\usage{
create_contact(client_id)
}
\arguments{
\item{client_id}{a client ID}
}
\value{
a single row tibble containing one contact episode
}
\description{
Creates a contact episode
}
\examples{
demo <- create_client(1)
create_contact(demo)
}
