% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{create_contacts}
\alias{create_contacts}
\title{Create a dataframe of contact episodes}
\usage{
create_contacts(df)
}
\arguments{
\item{df}{a data frame of consumer demographics containing a client_id}
}
\value{
a dataframe of service contact episodes where the count of episodes pre client is poisson distributed.
}
\description{
Create a dataframe of contact episodes
}
\examples{
demo <- create_client(1)
contact <- create_contacts(demo)
}
