% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{create_intakes}
\alias{create_intakes}
\title{Create a dataframe of intake forms}
\usage{
create_intakes(contacts)
}
\arguments{
\item{contacts}{a dataframe of service contacts}
}
\value{
a dataframe of intake forms
}
\description{
Create a dataframe of intake forms
}
\examples{
demo <- create_client(1)
contact <- create_contacts(demo)
contact <- localise_clients(contact)
create_intakes(contact)
}
