% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{create_k10s}
\alias{create_k10s}
\title{Creates a dataframe of entry and exit k10s}
\usage{
create_k10s(contacts, intakes)
}
\arguments{
\item{contacts}{a dataframe of contacts generated by create_contacts()}

\item{intakes}{a dataframe of contacts generated by create_intakes()}
}
\value{
a dataframe of entry and exit k10s
}
\description{
Creates a dataframe of entry and exit k10s
}
\examples{
demo <- create_client(1)
contact <- create_contacts(demo)
contact <- localise_clients(contact)
intakes <- create_intakes(contact)
create_k10s(contact, intakes)
}
