% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{localise_clients}
\alias{localise_clients}
\title{Adds a service contact location to a service contacts dataframe}
\usage{
localise_clients(contacts)
}
\arguments{
\item{contacts}{a dataframe of service contacts}
}
\value{
the input dataframe augmented with a service site
}
\description{
Adds a service contact location to a service contacts dataframe
}
\examples{
demo <- create_client(1)
contact <- create_contacts(demo)
contact <- localise_clients(contact)
}
