% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{ranger}
\alias{ranger}
\title{Put scores in range}
\usage{
ranger(x, decimals = 0, min, max)
}
\arguments{
\item{x}{numeric to put in range}

\item{decimals}{number of decimal places}

\item{min}{minimum value}

\item{max}{maximum value}
}
\value{
a numeric
}
\description{
Put scores in range
}
\examples{
ranger(3.4, decimals = 0, min = 1, max = 7)
ranger(0, decimals = 0, min = 1, max = 7)
}
